pyproject.toml
[project]
name = "a2a-udemy-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "sse-starlette>=3.0.3",
    "uvicorn>=0.38.0",
    "typer>=0.12.3",
    "ruff>=0.14.10",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "a2a-sdk[grpc]>=1.0.1",
]
